Unclosed braces in example
Braces in the example code has an unclosed brace `{{searchResult.entry.thumbnail.first().url` which causes the example code to throw an error. Craft does not give a clear error warning that this is the case and therefore is confusing for those who are new to Craft and/or this plugin.